CDSTP-86 Move LogicBlockHolder class to walk_puml_logic_graph.py (#78)
* CDSTP-86 Move LogicBlockHolder class to walk_puml_logic_graph.py

* CDSTP-86 Add class SubGraphNode extending class Node

* CDSTP-86 Add NodeTuple class

* CDSTP-86 Fix reference to create_puml_graph_from_node_class_graph

* CDSTP-86 Update SubGraphNode setter to use graph_data tuple

* CDSTP-86 - Fix linting error

* CDSTP-86 Update class docstrings

* CDSTP-86 Update SubGraphNode and NodeTuple in line with comments

* CDSTP-86 Migrate tests for LogicBlockHolder class to own file

* CDSTP-86 Fix linting

* CDSTP-86 Run mypy script

* CDSTP-86 Fix mypy error

* CDSTP-86 fix return error mypy
